+++ Runtime/Core/EaseDirection.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 
 namespace FlowTween {
-    [Flags]
-    public enum EaseDirection {
-        In = 1 << 0,
-        Out = 1 << 1,
-        InOut = In | Out
-    }
+
+[Flags]
+public enum EaseDirection {
+    In = 1 << 0,
+    Out = 1 << 1,
+    InOut = In | Out
+}
+
 }