+++ Runtime/Core/TweenBaseExtensions.cs
@@ -12,11 +12,6 @@ public static class TweenBaseExtensions {
     /// <summary>
     /// Sets the tween's <see cref="TweenBase.EaseFunction"/>.
     /// </summary>
-    /// <param name="ease">
-    /// The easing function. Input is the <see cref="TweenBase.RawProgress"/> of the tween,
-    /// between 0 and 1. The output can be outside of the 0-1 range, but it's recommended
-    /// to keep it close. The function should output 0 when the input is 0 and 1 when the input is 1.
-    /// </param>
     public static T Ease<T>(this T tween, Func<float, float> ease) where T : TweenBase {
         tween.EaseFunction = ease;
         return tween;
@@ -24,11 +19,42 @@ public static class TweenBaseExtensions {
     
     /// <summary>
     /// Sets the tween's <see cref="TweenBase.EaseFunction"/> to a function
-    /// representing the given <see cref="EaseType"/>.
+    /// represented by the given <see cref="Ease"/>.
     /// </summary>
-    public static T Ease<T>(this T tween, EaseType ease) where T : TweenBase {
-        tween.EaseFunction = EaseUtil.GetFunction(ease);
-        return tween;
+    public static T Ease<T>(this T tween, Ease ease) where T : TweenBase {
+        return tween.Ease(ease.GetFunction());
+    }
+    
+    /// <summary>
+    /// Sets the tween's <see cref="TweenBase.EaseFunction"/> to a function
+    /// represented by the given <see cref="EaseType"/> and <see cref="EaseDirection"/>.
+    /// </summary>
+    public static T Ease<T>(this T tween, EaseType type, EaseDirection direction) where T : TweenBase {
+        return tween.Ease(new Ease {
+            Type = type,
+            Direction = direction
+        });
+    }
+    
+    /// <summary>
+    /// Sets the tween to ease in with the given <see cref="EaseType"/>.
+    /// </summary>
+    public static T EaseIn<T>(this T tween, EaseType type) where T : TweenBase {
+        return tween.Ease(type, EaseDirection.In);
+    }
+    
+    /// <summary>
+    /// Sets the tween to ease out with the given <see cref="EaseType"/>.
+    /// </summary>
+    public static T EaseOut<T>(this T tween, EaseType type) where T : TweenBase {
+        return tween.Ease(type, EaseDirection.Out);
+    }
+    
+    /// <summary>
+    /// Sets the tween to ease in and out with the given <see cref="EaseType"/>.
+    /// </summary>
+    public static T EaseInOut<T>(this T tween, EaseType type) where T : TweenBase {
+        return tween.Ease(type, EaseDirection.InOut);
     }
     
     /// <summary>
@@ -36,17 +62,9 @@ public static class TweenBaseExtensions {
     /// The tween starts at <c>time=0</c> and ends at <c>time=1</c>.
     /// The curve's value is unconstrained, but it's recommended to keep it close to the 0-1 range.
     /// </summary>
-    public static T Ease<T>(this T tween, AnimationCurve ease) where T : TweenBase {
-        tween.EaseFunction = ease.Evaluate;
+    public static T Ease<T>(this T tween, AnimationCurve curve) where T : TweenBase {
+        tween.EaseFunction = curve.Evaluate;
         return tween;
-    }
-    
-    /// <summary>
-    /// Sets the tween's <see cref="TweenBase.CompleteAction"/>.
-    /// </summary>
-    public static T OnComplete<T>(this T runnable, Action action) where T : Runnable {
-        runnable.CompleteAction += action;
-        return runnable;
     }
     
     /// <summary>
@@ -72,28 +90,15 @@ public static class TweenBaseExtensions {
     }
     
     /// <summary>
-    /// Pauses the tween.
-    /// </summary>
-    /// <seealso cref="Runnable.IsPaused"/>
-    public static T Pause<T>(this T tween) where T : Runnable {
-        tween.IsPaused = true;
-        return tween;
-    }
-    
-    /// <summary>
-    /// Resumes the tween.
-    /// </summary>
-    /// <seealso cref="Runnable.IsPaused"/>
-    public static T Resume<T>(this T tween) where T : Runnable {
-        tween.IsPaused = false;
-        return tween;
-    }
-    
-    /// <summary>
     /// Sets the tween's <see cref="TweenBase.Duration"/>.
     /// </summary>
     public static T SetDuration<T>(this T tween, float duration) where T : TweenBase {
-        tween.Duration = duration;
+        tween.SetDurationInternal(duration);
+        return tween;
+    }
+    
+    public static T SetDelay<T>(this T tween, float delay) where T : TweenBase {
+        tween.Delay = delay;
         return tween;
     }