+++ Runtime/Core/TweenSettings.cs
@@ -11,8 +11,10 @@ namespace FlowTween {
 public class TweenSettings {
     [Min(0)]
     [SerializeField] float _duration = 1;
+    [Min(0)]
+    [SerializeField] float _delay;
     [SerializeField] EasingType _easeType;
-    [SerializeField] EaseType _presetEase;
+    [SerializeField] Ease _presetEase;
     [SerializeField] AnimationCurve _customEase;
     [SerializeField] LoopMode _loopMode;
 
@@ -23,12 +25,20 @@ public class TweenSettings {
         get => _duration;
         set => _duration = Mathf.Max(0, value);
     }
+    
+    /// <summary>
+    /// Delay in seconds, must be greater than 0.
+    /// </summary>
+    public float Delay {
+        get => _delay;
+        set => _delay = Mathf.Max(0, value);
+    }
 
     /// <summary>
     /// The preset easing type to use.
     /// Setting this also sets <see cref="EaseType"/> to <see cref="EasingType.Preset"/>.
     /// </summary>  
-    public EaseType PresetEase {
+    public Ease PresetEase {
         get => _presetEase;
         set {
             _presetEase = value;
@@ -80,7 +90,10 @@ public class TweenSettings {
     /// <see cref="EaseType"/> is not a valid <see cref="EasingType"/> value
     /// </exception>
     public void Apply(TweenBase tween) {
-        tween.SetDuration(_duration).Loop(_loopMode);
+        tween
+            .SetDuration(_duration)
+            .SetDelay(_delay)
+            .Loop(_loopMode);
 
         switch (_easeType) {
             case EasingType.Preset: